Moved final undisturbed T->T calculation block to numba_utilities.py.
