+++ src/LULUCF/scripts/core_model/LULUCF_fluxes.py
@@ -25,170 +25,6 @@ from ..utilities import universal_utilities as uu
 from ..utilities import log_utilities as lu
 from ..utilities import numba_utilities as nu
 
-import  math
-
-# Gross and net fluxes and ending carbon stocks for trees remaining trees with non-stand-replacing disturbances
-@jit(nopython=True)
-def calc_T_T_no_disturbs_local(node, most_recent_year_burned, agc_rf, bgc_rf, c_pools_fire_CO2, c_pools_fire_non_CO2,
-                               interval_end_year, c_dens_in,
-                               most_recent_year_not_tall_veg, Cf, Gef_co2, Gef_ch4, Gef_n2o,
-                               deadwood_c_ratio, litter_c_ratio):
-
-    # Retrieves the starting densities for each carbon pool from the input array
-    agc_dens_in, bgc_dens_in, deadwood_c_dens_in, litter_c_dens_in = nu.unpack_starting_carbon_densities(c_dens_in)
-
-    # Establishes which carbon pools are emitted if a fire was detected during the interval.
-    agc_ef_CO2, bgc_ef_CO2, deadwood_c_ef_CO2, litter_c_ef_CO2 = nu.unpack_emission_factors(c_pools_fire_CO2)
-
-
-    ## Step 1: Calculates the number of years of carbon gain before loss occurred
-    if most_recent_year_burned > 0:
-        # If a forest disturbance was detected, the gain_year_count are the number of years until detection of the last disturbance.
-        # There is no growth in the year of disturbance or the years after.
-        # The - 1 at the excludes the disturbance year from the gain_year_count since we decided there are no removals in the disturbance year.
-        # For example, if the time interval is 2010-2015 and the disturbance is detected in 2013 (t-2),
-        # there should be 2 years of growth (years t-4 and t-3, 2011 and 2012).
-        # This table illustrates each case for the example interval of 2010-2015.
-        # 0 years         11               - ((2015              - 2000)                - 5) - 1   (year t-4)
-        # 1 years         12               - ((2015              - 2000)                - 5) - 1   (year t-3)
-        # 2 years         13               - ((2015              - 2000)                - 5) - 1   (year t-2)
-        # 3 years         14               - ((2015              - 2000)                - 5) - 1   (year t-1)
-        # 4 years         15               - ((2015              - 2000)                - 5) - 1   (year t)
-        gain_year_count = most_recent_year_burned - ((interval_end_year - cn.first_model_year) - cn.interval_years) - 1
-    else:
-        gain_year_count = cn.interval_years
-
-
-    # Step 2: Assigns deadwood C and litter C ratios for removal factors, if relevant.
-    # Deadwood and litter C removals only occur in pixels that were not tall vegetation at some point (natural forest only).
-    # Thus, we need to check whether the pixel was non-tall vegetation at some point during the model before the end of this interval.
-    if (most_recent_year_not_tall_veg >= cn.first_model_year) or (most_recent_year_not_tall_veg == interval_end_year) or (most_recent_year_not_tall_veg == 0):
-        deadwood_c_ratio = 0.0
-        litter_c_ratio = 0.0
-
-
-    # Step 3: Calculates gross removals by carbon pools before disturbance. Gross removals are negative.
-    agc_gross_removals_out = float((agc_rf * gain_year_count) * -1) #float() necessary for Numba typing
-    bgc_gross_removals_out = float((bgc_rf * gain_year_count) * -1) #float() necessary for Numba typing
-    deadwood_c_gross_removals_out= agc_gross_removals_out * deadwood_c_ratio
-    litter_c_gross_removals_out= agc_gross_removals_out * litter_c_ratio
-
-    # Consolidates outputs into arrays to reduce the number of arguments returned to the decision tree.
-    # Must specify float32 because numba is quite particular about datatypes.
-    c_gross_removals_out = np.array([agc_gross_removals_out, bgc_gross_removals_out, deadwood_c_gross_removals_out, litter_c_gross_removals_out]).astype('float32')
-
-
-    # Step 4: Calculates carbon densities at the year of loss by carbon pool
-    agc_pre_disturb = agc_dens_in - agc_gross_removals_out
-    bgc_pre_disturb = bgc_dens_in - bgc_gross_removals_out
-    deadwood_c_pre_disturb = deadwood_c_dens_in - deadwood_c_gross_removals_out
-    litter_c_pre_disturb = litter_c_dens_in - litter_c_gross_removals_out
-
-    # Pre-disturbance carbon densities as an array, used as input for non-CO2 fire emissions and post-disturbance removals (if applicable)
-    c_pre_disturb = np.array([agc_pre_disturb, bgc_pre_disturb, deadwood_c_pre_disturb, litter_c_pre_disturb])
-
-
-    # Step 5: Calculates CO2 gross emissions by carbon pools.  Which ones are emitted depends on whether fire was detected.
-
-    # Calculates CO2 emissions from fire for each C pool using fire emission factors
-    # if a Gef for CO2 is supplied AND if there was fire during the interval.
-    # This is used for non-stand replacing forest disturbances (as opposed to entire C pools being combusted).
-    # From IPCC 2019 Eqn. 2.27
-    if most_recent_year_burned > 0:
-
-        #TODO Do these equations actually equal gross emissions in Mg C/ha (that's what subsequent calcs are assuming)?
-        # I'm not sure the units are correct. Could be CO2/ha or something else.
-        agc_gross_emis_out = ((agc_pre_disturb / cn.biomass_to_carbon_non_mangrove) * Cf * Gef_co2 * cn.g_to_kg) / cn.C_to_CO2 * agc_ef_CO2
-        bgc_gross_emis_out = ((bgc_pre_disturb / cn.biomass_to_carbon_non_mangrove) * Cf * Gef_co2 * cn.g_to_kg) / cn.C_to_CO2 * bgc_ef_CO2
-        deadwood_c_gross_emis_out = ((deadwood_c_pre_disturb / cn.biomass_to_carbon_non_mangrove) * Cf * Gef_co2 * cn.g_to_kg) / cn.C_to_CO2 * deadwood_c_ef_CO2
-        litter_c_gross_emis_out = ((litter_c_pre_disturb / cn.biomass_to_carbon_non_mangrove) * Cf * Gef_co2 * cn.g_to_kg) / cn.C_to_CO2 * litter_c_ef_CO2
-
-        # # For testing CO2 fire emissions
-        # print("c_dens_in:", c_dens_in)
-        # print("agc_rf:", agc_rf)
-        # print("gain_year_count:", gain_year_count)
-        # print("agc_pre_disturb:", agc_pre_disturb)
-        # print(f"Cf: {Cf}; Gef_co2: {Gef_co2}")
-        # print("AGC emission factor for fire:", agc_ef_CO2)
-        # print("agc_gross_emis_out:", agc_gross_emis_out)
-        # os.quit()
-
-    # No emissions if no fire detected
-    else:
-
-        agc_gross_emis_out = 0
-        bgc_gross_emis_out = 0
-        deadwood_c_gross_emis_out = 0
-        litter_c_gross_emis_out = 0
-
-    # Gross emissions as an array
-    c_gross_emissions_out = np.array([agc_gross_emis_out, bgc_gross_emis_out, deadwood_c_gross_emis_out, litter_c_gross_emis_out]).astype('float32')
-
-
-    # Step 6: Updates gross removals to include post-fire gross removals, if applicable.
-    # gain year count is the number of years between the disturbance and the end of the interval.
-    # This uses the same RFs before and after the fire.
-    if most_recent_year_burned > 0:
-
-        post_dist_RF = np.array([agc_rf, bgc_rf, agc_rf*deadwood_c_ratio, agc_rf*litter_c_ratio]).astype('float32')
-        post_dist_gain_year_count = cn.interval_years - gain_year_count - 1
-        post_dist_gross_removals = post_dist_gain_year_count * post_dist_RF
-
-        c_gross_removals_out = c_gross_removals_out - post_dist_gross_removals
-
-        # print("post_dist_RF:", post_dist_RF)
-        # print("gain_year_count:", gain_year_count)
-        # print("post_dist_gain_year_count:", post_dist_gain_year_count)
-        # print("post_dist_gross_removals:", post_dist_gross_removals)
-        # print("c_gross_removals_out_after_dist:", c_gross_removals_out)
-        # os.quit()
-
-
-    # Step 7: Calculates ending carbon densities by carbon pool.
-    # Starts with carbon density in (list converted to np array), adds gross removals (subtracts negative value), subtracts emissions.
-    # Ending carbon pools are not affected by non-CO2 emissions in the next step.
-    c_dens_out = np.array(c_dens_in).astype('float32') - c_gross_removals_out - c_gross_emissions_out
-
-
-    # Step 7: Calculates non-CO2 emissions (if relevant)
-    # Default non-CO2 emissions values
-    ch4_flux_out = 0
-    n2o_flux_out = 0
-
-    # Only assigns fire node code and calculates CH4 and N2O emissions if the pixel burned in the last interval
-    if most_recent_year_burned > 0:
-
-        state_out = nu.accrete_node(node, 1)
-
-        # Selects just the carbon pools that have non-CO2 emissions from fire
-        c_pools_for_fire_non_CO2 = np.where(c_pools_fire_non_CO2 == 1, c_pre_disturb, 0)
-
-        # Sums the C pools that have non-CO2 fire emissions. We don't track which C pools the CH4 and N2O emissions come from,
-        # so the pools are combined.
-        c_pools_for_fire_total = np.sum(c_pools_for_fire_non_CO2)
-
-        # Calculates non-CO2 fire emissions using the selected C pools in the year before disturbance
-        ch4_flux_out, n2o_flux_out = nu.non_CO2_fire_equations(c_pools_for_fire_total, Cf, Gef_ch4, Gef_n2o)
-
-        # # For testing non-CO2 emissions
-        # print("c_dens_in:", c_dens_in)
-        # print("c_pre_disturb:", c_pre_disturb)
-        # print(f"Cf: {Cf}; Gef_ch4: {Gef_ch4}; GWP CH4: {cn.gwp_ch4}")
-        # print(f"Cf: {Cf}; Gef_n2o: {Gef_n2o}; GWP N2O: {cn.gwp_n2o}")
-        # print("c_pools_for_fire_non_CO2:", c_pools_for_fire_non_CO2)
-        # print("c_pools_for_fire_total:", c_pools_for_fire_total)
-        # print(f"ch4_flux_out: {ch4_flux_out}; n2o_flux_out: {n2o_flux_out};")
-        # os.quit()
-
-    # Node code if no fire in the last interval. No CH4 and N2O emissions calculated.
-    else:
-
-        state_out = nu.accrete_node(node, 2)
-
-    non_co2_fluxes_out = np.array([ch4_flux_out, n2o_flux_out]).astype('float32')
-
-    return state_out, c_gross_emissions_out, c_gross_removals_out, non_co2_fluxes_out, c_dens_out, gain_year_count
-
 
 # Function to calculate LULUCF fluxes and carbon densities
 # Operates pixel by pixel, so uses numba (Python compiled to C++).
